refactor(native-agent): extract terminal-command helpers
Move the harness-agnostic pieces of building a native terminal command
(review CLI PATH shim, companion module resolution, hook client command,
shell quoting, inline TOML) out of NativeReviewTurnLauncher into
terminal-command.ts, and make threads-cli import the shared env-var names
instead of redeclaring them. No behavior change.

Agent-Session: ba208f4e-0500-4cca-bc8c-67aafcc19811
Agent-Session: 01a0600c-8adb-7663-ba10-a0207d0507b8
Agent-Session: 01a0602c-1d0f-7b03-8356-7a3614e78dc3
Agent-Session: 01a06048-aee8-7142-84b8-ef4838717f58
Agent-Session: 01a06050-6825-7661-94cc-512afb8a405e
Agent-Session: 01a060a7-1371-7d61-8a00-a54bee4571bd
